requirements.txt
Django==4.2.21
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-environ==0.11.2
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
django-storages==1.14.2
moviepy==1.0.3
openai-whisper==20231117
google-generativeai==0.3.2
Pillow==10.2.0
pytest==8.0.2
pytest-django==4.8.0
django-markdownify==0.9.5 